Changed sizeof(float) to uint16

+++ source/cheetah-euxfel/agipd_reader.cpp
@@ -223,8 +223,9 @@ void cAgipdReader::open(char *baseFilename){
 	}
 
 	// Allocate memory for data and masks
-	data = (float*) malloc(nn*sizeof(uint16_t));
-	mask = (uint16_t*) malloc(nn*sizeof(uint16_t));
+	data = (float*) malloc(nn*sizeof(float));
+
+	mask = (uint16_t*) malloc(nn*sizeof(float));
 	digitalGain = (uint16_t*) malloc(nn*sizeof(uint16_t));
